Fixed pre-commit hook
