+++ .gitignore
@@ -4,3 +4,5 @@ backend/.env
 
 frontend/node_modules
 frontend/.env
+
+node_modules
