+++ .husky/pre-commit
@@ -1,0 +1,16 @@
+# npm test
+
+# echo "Running backend lint..."
+# cd backend
+# npm run lint
+
+# echo "Running frontend lint..."
+# cd..
+# cd frontend
+# npm run lint
+
+echo "Running backend lint..."
+npm --prefix backend run lint
+
+echo "Running frontend lint..."
+npm --prefix frontend run lint
