+++ backend/controllers/companyController.js
@@ -2,7 +2,7 @@ const companyService = require('../services/companyService');
 
 const companyController = {};
 
-companyController.getUsers = async (req, res) => {
+companyController.getUsers = async (req) => {
     const { companyId } = req.params;
     const companyUsers = await companyService.getUsers(companyId);
     return {
