+++ backend/controllers/userController.js
@@ -2,7 +2,7 @@ const userService = require('../services/userService');
 
 const userController = {};
 
-userController.getUsers = async (req, res) => {
+userController.getUsers = async () => {
     const users = await userService.getUsers();
 
     return {
