+++ backend/eslint.config.mjs
@@ -10,4 +10,15 @@ export default defineConfig([
     languageOptions: { globals: globals.node },
   },
   { files: ['**/*.js'], languageOptions: { sourceType: 'commonjs' } },
+  {
+    rules: {
+      "no-unused-vars": [
+        "error",
+        { 
+          "args": "all", 
+          "argsIgnorePattern": "^_" 
+        }
+      ]
+    }
+  }
 ]);
