+++ backend/services/companyService.js
@@ -1,4 +1,3 @@
-const { where } = require('sequelize');
 const db = require('../models');
 const UserModel = db.User;
