+++ frontend/eslint.config.js
@@ -17,5 +17,16 @@ export default defineConfig([
       globals: globals.browser,
       parserOptions: { ecmaFeatures: { jsx: true } },
     },
+    rules: {
+        "no-unused-vars": [
+          "error",
+          { 
+            "args": "all", 
+            "argsIgnorePattern": "^_",
+            "caughtErrors": "all",
+            "caughtErrorsIgnorePattern": "^_"
+          }
+        ]
+      }
   },
 ])
