+++ frontend/src/App.jsx
@@ -1,16 +1,15 @@
-import { useState } from 'react'
-import { BrowserRouter, Routes, Route, Navigate, RouterProvider } from 'react-router-dom'
-import { createBrowserRouter } from 'react-router-dom'
+// import { BrowserRouter, Routes, Route, Navigate, RouterProvider } from 'react-router-dom'
+// import { createBrowserRouter } from 'react-router-dom'
 
-import reactLogo from './assets/react.svg'
-import viteLogo from './assets/vite.svg'
-import heroImg from './assets/hero.png'
-import Dashboard from './pages/Dashboard'
+// import reactLogo from './assets/react.svg'
+// import viteLogo from './assets/vite.svg'
+// import heroImg from './assets/hero.png'
+// import Dashboard from './pages/Dashboard'
 // import Dashboard from './pages/Dashboard2'
 // import './App.css'
-import Login from './pages/Login'
-import Signup from './pages/Signup'
-import Home from './pages/Home'
+// import Login from './pages/Login'
+// import Signup from './pages/Signup'
+// import Home from './pages/Home'
 import AppRoute from './routes/AppRoute'
 
 // const router = createBrowserRouter([
