+++ frontend/src/compoenets/ProtectedRoute.jsx
@@ -1,10 +1,10 @@
 import { Navigate } from "react-router-dom";
 import { isLoggedIn } from "../util/auth";
 import { useContext } from "react";
-import { AuthContext } from "../context/AuthContext";
+import { AuthContextProvider} from "../context/AuthContextProvider";
 
 export default function ProtectedRoute({ children }) {
-    const AuthContextData = useContext(AuthContext);
+    const AuthContextData = useContext(AuthContextProvider);
     console.log('Protected Route AuthContextData: ',AuthContextData);
     if(AuthContextData.authLoading){
         return null;
