+++ frontend/src/pages/Home.jsx
@@ -1,11 +1,10 @@
 import { Link, NavLink } from "react-router-dom";
 import "../assets/css/home.css";
-import { isLoggedIn } from "../util/auth";
 import { useContext } from "react";
-import { AuthContext } from "../context/AuthContext";
+import { AuthContextProvider} from "../context/AuthContextProvider";
 
 function Home() {
-  const AuthContextData = useContext(AuthContext);
+  const AuthContextData = useContext(AuthContextProvider);
   const isLogged = AuthContextData.isLoggedIn;
   return (
     <div className="home-page">
