+++ frontend/src/pages/Login.jsx
@@ -3,10 +3,10 @@ import { useContext, useState } from "react";
 import "../assets/css/auth.css";
 import api from "../services/api";
 import { useNavigate, NavLink } from "react-router-dom";
-import { AuthContext } from "../context/AuthContext";
+import { AuthContextProvider} from "../context/AuthContextProvider";
 
 export default function Login() {
-  const AuthContextData = useContext(AuthContext);
+  const AuthContextData = useContext(AuthContextProvider);
   const navigate = useNavigate();
   const [apiError, setApiError] = useState("");
     const [email, setEmail] = useState("");
